requirements.txt
flask
python-dotenv
requests
beautifulsoup4
lxml
apscheduler
python-telegram-bot==20.3
selenium
sqlite3
